Migrer dernières notifications vers UserNotifier
